+++ packages/benchmark-js/src/Benchmark.ts
@@ -1,19 +1,20 @@
 import { tTable } from './constants';
-import { Logger, LoggerLevel } from './Logger';
+import { BenchmarkLoggerLevel, Logger } from './Logger';
 import { TU } from './TimeUnit';
+import { formatNumber, genStr, getMean, getMinTime, getVariance, sleep } from './tools';
 import { BenchmarkOptions, TestFn } from './types';
 import { BenchmarkStats, _BenchmarkSettings, _Nanosecond } from './types.internal';
-import { formatNumber, genStr, getMean, getMinTime, getVariance, hrtime2ns, sleep } from './utils';
 
 export class Benchmark {
     private static minTime: _Nanosecond = TU.ns(0);
 
-    private name: string;
+    private logger: Logger;
 
+    private name: string;
     private testFn: TestFn;
 
-    private onComplete: Optional<Function>;
-    private onStart: Optional<Function>;
+    private onComplete: Optional<() => void>;
+    private onStart: Optional<() => void>;
 
     private settings: Required<_BenchmarkSettings>;
 
@@ -43,36 +44,63 @@ export class Benchmark {
      * @param options The options of benchmark.
      */
     constructor(name: string, testFn: TestFn, options?: BenchmarkOptions) {
-        this.name = name;
-
-        this.testFn = testFn;
-
-        this.onComplete = options?.onComplete ?? null;
-        this.onStart = options?.onStart ?? null;
-
         if (Benchmark.minTime === 0) {
             Benchmark.minTime = TU.ns(Math.max(getMinTime() * 100, 50_000_000));
         }
 
+        this.logger = new Logger(name);
+
+        this.name = name;
+        this.testFn = testFn;
+
+        const {
+            delay = 5,
+            initCount = 1,
+            maxPrepareTime = 1_000,
+            maxTime = 5_000,
+            minSamples = 5,
+            minTime = 0,
+            onComplete = null,
+            onStart = null,
+        } = options ?? {};
+
+        this.onComplete = onComplete;
+        this.onStart = onStart;
+
         this.settings = {
-            delay: TU.ms2ns(options?.delay ?? 5),
-            initCount: options?.initCount ?? 1,
-            maxPrepareTime: TU.ms2ns(options?.maxPrepareTime ?? 1_000),
-            maxTime: TU.ms2ns(options?.maxTime ?? 5_000),
-            minSamples: options?.minSamples ?? 5,
-            minTime:
-                options?.minTime === undefined || options.minTime === 0 ? Benchmark.minTime : TU.ms2ns(options.minTime),
+            delay: TU.ms2ns(delay),
+            initCount,
+            maxPrepareTime: TU.ms2ns(maxPrepareTime),
+            maxTime: TU.ms2ns(maxTime),
+            minSamples,
+            minTime: minTime === 0 ? Benchmark.minTime : TU.ms2ns(minTime),
         };
 
         this.count = this.settings.initCount;
+
+        this.logConfigs();
     }
 
-    public static get loggerLevel(): LoggerLevel {
+    public static get loggerLevel(): BenchmarkLoggerLevel {
         return Logger.level;
     }
 
-    public static set loggerLevel(level: LoggerLevel) {
+    public static set loggerLevel(level: BenchmarkLoggerLevel) {
         Logger.level = level;
+    }
+
+    private logConfigs() {
+        this.logger.debug(`delay           : ${this.logger.beautifyNumber(this.settings.delay)} ns`);
+        this.logger.debug(`initial count   : ${this.logger.beautifyNumber(this.settings.initCount)}`);
+        this.logger.debug(`max prepare time: ${this.logger.beautifyNumber(this.settings.maxPrepareTime)} ns`);
+        this.logger.debug(`max time        : ${this.logger.beautifyNumber(this.settings.maxTime)} ns`);
+        this.logger.debug(`min samples     : ${this.logger.beautifyNumber(this.settings.minSamples)}`);
+        this.logger.debug(`min time        : ${this.logger.beautifyNumber(this.settings.minTime)} ns`);
+
+        this.logger.debug(`${this.onComplete ? 'Has' : 'No'} callback \`onComplete\``);
+        this.logger.debug(`${this.onStart ? 'Has' : 'No'} callback \`onStart\``);
+
+        this.logger.debug();
     }
 
     /**
@@ -84,11 +112,14 @@ export class Benchmark {
 
         this.onStart?.();
 
-        this.benchmarking(this.settings.maxPrepareTime);
+        this.benchmarking(this.settings.maxPrepareTime, true);
         // Delete samples generated while pre-benchmarking.
         this.stats.sample = [];
         this.benchmarking(this.settings.maxTime);
         this.evaluate();
+
+        this.logger.debug(this.toString());
+        this.logger.debug();
 
         this.onComplete?.();
 
@@ -101,20 +132,34 @@ export class Benchmark {
             this.testFn();
         }
         const duration = process.hrtime(begin);
-        return hrtime2ns(duration);
+        return TU.hrtime2ns(duration);
     }
 
-    private benchmarking(ns: _Nanosecond): void {
+    private logCycleData() {
+        const { sample } = this.stats;
+        const len = sample.length;
+        this.logger.debug(`${len.toString().padStart(3)}> period: ${sample[len - 1]} ns`);
+    }
+
+    private logCountChanging(from: number, to: number) {
+        this.logger.debug(`count: ${from} -> ${to}`);
+    }
+
+    private benchmarking(ns: _Nanosecond, prepare?: boolean): void {
         let elapsed: _Nanosecond = TU.ns(0);
         while (true) {
             const used = this.cycle();
             const period = TU.ns(used / this.count);
             this.stats.sample.push(period);
 
+            if (!prepare) this.logCycleData();
+
             // Calculate how many more iterations it will take to achieve the `minTime`.
             // After pre-benchmarking stage, we should get a good count number.
             if (used < this.settings.minTime) {
-                this.count += Math.ceil((this.settings.minTime - used) / period);
+                const count = this.count + Math.ceil((this.settings.minTime - used) / period);
+                this.logCountChanging(this.count, count);
+                this.count = count;
             }
 
             elapsed = TU.ns(elapsed + used);
@@ -162,10 +207,10 @@ export class Benchmark {
         const opsStr = formatNumber(this.stats.ops.toFixed(this.stats.ops < 100 ? 2 : 0));
         const rmeStr = this.stats.rme.toFixed(2);
 
-        return this.name + genStr(` ${opsStr} ops/sec`, ` ${rmeStr}%`, ` (${size} sample${size > 1 ? 's' : ''})`);
+        return genStr(`${opsStr} ops/sec`, ` ${rmeStr}%`, ` (${size} sample${size > 1 ? 's' : ''})`);
     }
 
-    public printResult(): void {
-        Logger.log(this.toString());
+    public writeResult(): void {
+        this.logger.write(this.toString());
     }
 }